chore: refactor merge-into static filter
