+++ scripts/ci/deploy/config/databend-query-node-1.toml
@@ -70,7 +70,7 @@ auth_type = "no_password"
 [log]
 
 [log.file]
-level = "DEBUG"
+level = "INFO"
 format = "text"
 dir = "./.databend/logs_1"
