+++ src/query/service/src/interpreters/mod.rs
@@ -43,6 +43,7 @@ mod interpreter_index_refresh;
 mod interpreter_insert;
 mod interpreter_kill;
 mod interpreter_merge_into;
+mod interpreter_merge_into_static_filter;
 mod interpreter_metrics;
 mod interpreter_network_policies_show;
 mod interpreter_network_policy_alter;
